0:44 hotfix for deleting

+++ src/js/main.js
@@ -352,27 +352,32 @@ function AppViewModel() {
 
 
     self.removeCapability = function (capability) {
+
         let unmapped = ko.mapping.toJS(viewModel.devices);
         console.log("remove from component " + unmapped[openedIndex].componentId + " the capability assertion id " + unmapped[openedIndex].capabilityAssertionId + " with the variant id " + capability.id);
 
+        //TODO: do only if button is not disabled, this is just a 0:44 hotfix
+        if (unmapped[openedIndex].componentId === "_jJdx4DD7EeiuBvcKgWzd3Q"){
+            console.log("deleted");
+            $.ajax({
+                url: viewModel.restConfig.hostname() + "/services/resourceinstance/" + unmapped[openedIndex].componentId + "/capability/" + unmapped[openedIndex].capabilityAssertionId + "/variant/" + capability.id,
+                type: "DELETE",
+                contentType: "application/json",
+                success: function () {
+                    //GUI updates
+                    //remove capability from devices
+                    unmapped[openedIndex].capability = $.grep(unmapped[openedIndex].capability, function (e) {
+                        return e.id !== capability.id;
+                    });
+                    ko.mapping.fromJS(unmapped, viewModel.devices);
 
-        $.ajax({
-            url: viewModel.restConfig.hostname() + "/services/resourceinstance/" + unmapped[openedIndex].componentId + "/capability/" + unmapped[openedIndex].capabilityAssertionId + "/variant/" + capability.id,
-            type: "DELETE",
-            contentType: "application/json",
-            success: function () {
-                //GUI updates
-                //remove capability from devices
-                unmapped[openedIndex].capability = $.grep(unmapped[openedIndex].capability, function (e) {
-                    return e.id !== capability.id;
-                });
-                ko.mapping.fromJS(unmapped, viewModel.devices);
+                    //remove capability from modal
+                    viewModel.currentCapability(unmapped[openedIndex].capability)
 
-                //remove capability from modal
-                viewModel.currentCapability(unmapped[openedIndex].capability)
+                }
+            });
+        }
 
-            }
-        });
     };
